Branching.

+++ EnchantmentsEnhance-Core/src/main/java/org/pixeltime/enchantmentsenhance/gui/menu/MainMenu.java
@@ -22,8 +22,10 @@ import org.bukkit.Bukkit;
 import org.bukkit.DyeColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
+import org.bukkit.event.inventory.ClickType;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.scheduler.BukkitRunnable;
+import org.bukkit.scheduler.BukkitTask;
 import org.pixeltime.enchantmentsenhance.Main;
 import org.pixeltime.enchantmentsenhance.event.Enhance;
 import org.pixeltime.enchantmentsenhance.gui.Clickable;
@@ -37,12 +39,30 @@ import org.pixeltime.enchantmentsenhance.util.ItemBuilder;
 import org.pixeltime.enchantmentsenhance.util.Util;
 import org.pixeltime.enchantmentsenhance.util.XMaterial;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 public class MainMenu extends GUIAbstract {
+    public static final ItemStack[] WOOL = {
+            new ItemBuilder(XMaterial.BLACK_WOOL.parseMaterial()).setDyeColor(DyeColor.BLACK).toItemStack(),
+            new ItemBuilder(XMaterial.BLUE_WOOL.parseMaterial()).setDyeColor(DyeColor.BLUE).toItemStack(),
+            new ItemBuilder(XMaterial.BROWN_WOOL.parseMaterial()).setDyeColor(DyeColor.BROWN).toItemStack(),
+            new ItemBuilder(XMaterial.CYAN_WOOL.parseMaterial()).setDyeColor(DyeColor.CYAN).toItemStack(),
+            new ItemBuilder(XMaterial.LIGHT_GRAY_WOOL.parseMaterial()).setDyeColor(DyeColor.LIGHT_GRAY).toItemStack(),
+            new ItemBuilder(XMaterial.GREEN_WOOL.parseMaterial()).setDyeColor(DyeColor.GREEN).toItemStack(),
+            new ItemBuilder(XMaterial.LIGHT_BLUE_WOOL.parseMaterial()).setDyeColor(DyeColor.LIGHT_BLUE).toItemStack(),
+            new ItemBuilder(XMaterial.GRAY_WOOL.parseMaterial()).setDyeColor(DyeColor.GRAY).toItemStack(),
+            new ItemBuilder(XMaterial.LIME_WOOL.parseMaterial()).setDyeColor(DyeColor.LIME).toItemStack(),
+            new ItemBuilder(XMaterial.MAGENTA_WOOL.parseMaterial()).setDyeColor(DyeColor.MAGENTA).toItemStack(),
+            new ItemBuilder(XMaterial.ORANGE_WOOL.parseMaterial()).setDyeColor(DyeColor.ORANGE).toItemStack(),
+            new ItemBuilder(XMaterial.PINK_WOOL.parseMaterial()).setDyeColor(DyeColor.PINK).toItemStack(),
+            new ItemBuilder(XMaterial.PURPLE_WOOL.parseMaterial()).setDyeColor(DyeColor.PURPLE).toItemStack(),
+            new ItemBuilder(XMaterial.RED_WOOL.parseMaterial()).setDyeColor(DyeColor.RED).toItemStack(),
+            new ItemBuilder(XMaterial.WHITE_WOOL.parseMaterial()).setDyeColor(DyeColor.WHITE).toItemStack(),
+            new ItemBuilder(XMaterial.YELLOW_WOOL.parseMaterial()).setDyeColor(DyeColor.YELLOW).toItemStack()
+    };
     public static Map<String, ItemStack> itemOnEnhancingSlot = new HashMap<String, ItemStack>();
-    public static Map<String, Clickable> enhanceMode = new HashMap<>();
+    public static Map<String, Clickable> enhancingMode = new HashMap<>();
+    public static Map<String, BukkitTask> inProgress = new HashMap<>();
     public static EnhanceIcon enhance = new EnhanceIcon();
     public static ForceIcon force = new ForceIcon();
     public static RemoveIcon remove = new RemoveIcon();
@@ -54,14 +74,20 @@ public class MainMenu extends GUIAbstract {
     public static GearIcon gear = new GearIcon();
     public static ToolIcon tool = new ToolIcon();
     public static AccessoryIcon accessory = new AccessoryIcon();
+    public static CancelIcon cancel = new CancelIcon();
 
     public MainMenu(Player player) {
-        super(player, 54, SettingsManager.lang.getString("Menu.gui.title"));
+        super(player, 54, SettingsManager.lang.getString("menu.gui.title"));
         update();
     }
 
     public static void clearPlayer(String playerName) {
         itemOnEnhancingSlot.remove(playerName);
+    }
+
+    public static ItemStack randomWool() {
+        Random random = new Random();
+        return WOOL[random.nextInt(WOOL.length)];
     }
 
     @Override
@@ -70,25 +96,91 @@ public class MainMenu extends GUIAbstract {
         getActions().clear();
         Player player = Bukkit.getPlayer(playerName);
 
-        if (!enhanceMode.containsKey(playerName)) {
-            enhanceMode.put(playerName, gear);
+        if (!enhancingMode.containsKey(playerName)) {
+            enhancingMode.put(playerName, gear);
         }
 
         if (itemOnEnhancingSlot.containsKey(playerName)) {
+            // Sets Enhancing item on display.
             setItem(Util.getSlot(8, 4), itemOnEnhancingSlot.get(playerName));
 
-            setItem(enhance.getPosition(), enhance.getItem(itemOnEnhancingSlot.get(playerName)), (clickType) ->
-                    Enhance.diceToEnhancement(itemOnEnhancingSlot.get(playerName), player));
+            if (Enhance.getValidationOfItem(itemOnEnhancingSlot.get(playerName)) && Enhance.getValidationOfPlayer(itemOnEnhancingSlot.get(playerName), player)) {
+                setItem(enhance.getPosition(), enhance.getGlowingItem(itemOnEnhancingSlot.get(playerName)), (clickType) -> {
+                    if (clickType == ClickType.LEFT || clickType == ClickType.DOUBLE_CLICK) {
+                        inProgress.put(playerName, new BukkitRunnable() {
+                            int count = 0;
+                            List<ItemStack> animate = new ArrayList<>();
+
+                            @Override
+                            public void run() {
+                                if (count == 0) {
+                                    setItem(cancel.getPosition(), cancel.getGlowingItem(playerName), (clicktype) -> {
+                                        this.cancel();
+                                        update();
+                                    });
+                                }
+                                if (count == 5) {
+                                    try {
+                                        Enhance.diceToEnhancement(itemOnEnhancingSlot.get(playerName), player);
+                                    } catch (Exception ex) {
+                                        // Player might not be online.
+                                    }
+                                    this.cancel();
+                                    update();
+                                    inProgress.remove(playerName);
+                                    return;
+                                }
+                                animate.add(randomWool());
+                                if (count >= 0) {
+                                    for (int i = 0; i < animate.size(); i++) {
+                                        setItem(Util.getSlot(3 + i, 4), animate.get(animate.size() - i - 1));
+                                    }
+                                }
+                                count++;
+                            }
+                        }.runTaskTimer(Main.getMain(), 0L, 10L));
+                    } else {
+                        // Right click.
+                        Enhance.diceToEnhancement(itemOnEnhancingSlot.get(playerName), player);
+                    }
+                });
+            } else {
+                setItem(enhance.getPosition(), enhance.getItem(itemOnEnhancingSlot.get(playerName)), (clickType) -> {
+                    int enchantLevel = ItemManager.getItemEnchantLevel(itemOnEnhancingSlot.get(playerName)) + 1;
+                    int stoneId = Enhance.getStoneId(itemOnEnhancingSlot.get(playerName), enchantLevel);
+                    Util.sendMessage(SettingsManager.lang.getString("item.noItem")
+                            .replaceAll("%STONE%", SettingsManager.lang.getString(
+                                    "item." + stoneId)), player);
+                });
+            }
 
 
             if (DataManager.maximumFailstackApplied[ItemManager.getItemEnchantLevel(itemOnEnhancingSlot.get(playerName)) + 1] > 0
                     && DataManager.costToForceEnchant[ItemManager.getItemEnchantLevel(itemOnEnhancingSlot.get(playerName)) + 1] > 0) {
-                setItem(force.getPosition(), force.getItem(itemOnEnhancingSlot.get(playerName)), (clickType) ->
-                        Enhance.forceToEnhancement(itemOnEnhancingSlot.get(playerName), player));
+                if (Enhance.getValidationOfForce(itemOnEnhancingSlot.get(playerName), player)) {
+                    setItem(force.getPosition(), force.getGlowingItem(itemOnEnhancingSlot.get(playerName)), (clickType) -> {
+                        Enhance.forceToEnhancement(itemOnEnhancingSlot.get(playerName), player);
+                    });
+                } else {
+                    setItem(force.getPosition(), force.getItem(itemOnEnhancingSlot.get(playerName)), (clickType) -> {
+                        // Current enchant level before enhancing
+                        int enchantLevel = ItemManager.getItemEnchantLevel(itemOnEnhancingSlot.get(playerName)) + 1;
+                        // Finds the stone used in the enhancement
+                        int stoneId = Enhance.getStoneId(itemOnEnhancingSlot.get(playerName), enchantLevel);
+                        Util.sendMessage(SettingsManager.lang.getString("item.noItem")
+                                .replaceAll("%STONE%", SettingsManager.lang.getString(
+                                        "item." + stoneId)), player);
+                    });
+                }
             }
 
-            setItem(remove.getPosition(), remove.getGlowingItem(playerName), (clickType) ->
-                    clearPlayer(playerName));
+            setItem(remove.getPosition(), remove.getGlowingItem(playerName), (clickType) -> {
+                clearPlayer(playerName);
+                if (MainMenu.inProgress.containsKey(player.getName())) {
+                    MainMenu.inProgress.get(player.getName()).cancel();
+                }
+            });
+
 
             setItem(stats.getPosition(), stats.getItem(playerName));
 
@@ -123,9 +215,19 @@ public class MainMenu extends GUIAbstract {
                     }
                 }.runTaskLaterAsynchronously(Main.getMain(), 2L));
 
-        setItem(gear.getPosition(), gear.getItem(playerName));
-        setItem(tool.getPosition(), tool.getItem(playerName));
-        setItem(accessory.getPosition(), accessory.getItem(playerName));
+        setItem(gear.getPosition(), (enhancingMode.containsKey(playerName) && enhancingMode.get(playerName).equals(gear))
+                        ? gear.getGlowingItem(playerName)
+                        : gear.getItem(playerName),
+                (clickType) -> enhancingMode.put(playerName, gear));
+
+        setItem(tool.getPosition(), (enhancingMode.containsKey(playerName) && enhancingMode.get(playerName).equals(tool))
+                        ? tool.getGlowingItem(playerName)
+                        : tool.getItem(playerName),
+                (clickType) -> enhancingMode.put(playerName, tool));
+        setItem(accessory.getPosition(), (enhancingMode.containsKey(playerName) && enhancingMode.get(playerName).equals(accessory))
+                        ? accessory.getGlowingItem(playerName)
+                        : accessory.getItem(playerName),
+                (clickType) -> enhancingMode.put(playerName, accessory));
 
         for (int i : MenuCoord.getPlaceHolderCoords()) {
             setItem(i, new ItemBuilder(XMaterial.BLACK_STAINED_GLASS_PANE.parseMaterial()).setDyeColor(DyeColor.BLACK).setName("&0").toItemStack());
